day 9 part 1

+++ c#/day 8 part 1/Program.cs
@@ -30,4 +30,4 @@ while(currentNode != "ZZZ")
     result++;
 }
 
-Console.WriteLine((long)result);
+Console.WriteLine(result);
